Holi pipeline
This script is a pipeline for processing and aligning NGS dataset against a user defined database. The pipeline is a shell script which needs to be saved locally but not installed. Please find the lists of dependent programmes at the start of the script.

For comments, suggestions and feedback please contact Mikkel Winther Pedersen (mwpedersen@snm.ku.dk). This pipeline has been submitted for publishing but is not yet published.  

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.